+++ src/main/java/net/more_rpg_classes/custom/spell_impacts/RushForwardToTargetSpellImpact.java
@@ -30,6 +30,7 @@ public class RushForwardToTargetSpellImpact implements SpellHandlers.CustomImpac
                 Vec3d velocity = direction.normalize().multiply(distance);
                 caster.setVelocity(velocity);
                 caster.velocityModified = true;
+                caster.velocityDirty = true;
             }
         }